fixings

+++ app/assets/javascripts/app.js.coffee
@@ -1,7 +1,5 @@
 # Define App and dependencies
-BookApp = angular.module("BookApp", [
-  "ngRoute"
-])
+BookApp = angular.module("BookApp", ["ngRoute"])
 
 # Setup the angular router
 BookApp.config ["$routeProvider", "$locationProvider", ($routeProvider, $locationProvider)->
@@ -18,7 +16,7 @@ BookApp.config ["$routeProvider", "$locationProvider", ($routeProvider, $locatio
   $locationProvider.html5Mode(true)
 ]
 
-# Books Contrrails oller
+# Books Controller
 BookApp.controller "BooksCtrl", ["$scope", "$http", ($scope, $http) ->
 
   $scope.books = []
@@ -28,20 +26,18 @@ BookApp.controller "BooksCtrl", ["$scope", "$http", ($scope, $http) ->
       $scope.books = data
 
   $scope.addBook = ->
-    console.log $scope.newBook
     $http.post("/books.json", $scope.newBook).success (data)->
-      console.log "BOOK SAVED!"
       $scope.newBook = {}
       $scope.books.push(data)
 
   $scope.deleteBook = ->
-    console.log @book
-    $http.delete("/books/#{@book.id}.json").success (data)=>
-      console.log "book deleted"
-      $scope.books.splice(@$index,1)
+    conf = confirm "Are you sure?"
+    if conf
+      $http.delete("/books/#{@book.id}.json").success (data)=>
+        $scope.books.splice(@$index,1)
 ]
 
-# BookDetailsCtrl
+# BookDetails Controller
 BookApp.controller "BookDetailsCtrl", ["$scope", "$http", "$routeParams", ($scope, $http, $routeParams)->
   $scope.book_id = $routeParams.id
 
@@ -60,7 +56,7 @@ BookApp.controller "BookDetailsCtrl", ["$scope", "$http", "$routeParams", ($scop
     $http.put("/books/#{$scope.book_id}.json", $scope.book).success (data)->
 ]
 
-# Define Config
+# Define Config for CSRF token
 BookApp.config(["$httpProvider", ($httpProvider)->
   $httpProvider.defaults.headers.common['X-CSRF-Token'] = $('meta[name=csrf-token]').attr('content')
 ])
